Edit: adjusted weight as decimal

+++ database/migrations/2023_10_03_141142_create_movements_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name')->unique();
             $table->unsignedTinyInteger('reps')->nullable();
-            $table->unsignedSmallInteger('max_weight')->nullable();
+            $table->decimal('max_weight', 4, 1)->nullable();
             $table->unsignedBigInteger('user_id');
             $table->foreign('user_id')
                     ->references('id')
